+++ src/ImageSharp.Drawing/Processing/ShapeOptions.cs
@@ -26,7 +26,7 @@ namespace SixLabors.ImageSharp.Drawing.Processing
         /// </summary>
         public IntersectionRule IntersectionRule { get; set; } = IntersectionRule.OddEven;
 
-        public bool UsePolygonScanner { get; set; } = false;
+        public bool UsePolygonScanner { get; set; } = true;
 
         /// <inheritdoc/>
         public ShapeOptions DeepClone() => new ShapeOptions(this);
